add balanced classifier, diff tailed

+++ balanced-classifier/balanced-classifier.py
@@ -1,7 +1,8 @@
-from pyexpat import model
+# from pyexpat import model
 import torch 
 import torch.nn as nn
 import torch.nn.functional as F
+from os import path
 from torch.utils.data import Dataset
 from torchvision import datasets, transforms, models
 from resnet32 import resnet32
@@ -13,7 +14,7 @@ import math
 from collections import Counter
 wandb.login(key= "3ab0ccc1e73901e7d78c5af7de65194191805602")
 wandb.init(project="balanced classifier", name="finetune", config={"num_epochs":200, "batch_size":128,
-         "lr":0.1, "momentum":0.9, "weight_decay":0.0001, "num_epochs_fine":25}) # set mode="disabled" to not track logs
+         "lr":0.1, "momentum":0.9, "weight_decay":0.0001, "num_epochs_fine":25}, mode="disabled") # set mode="disabled" to not track logs
 config = wandb.config
 
 
@@ -22,19 +23,25 @@ def make_lt_dataset(train_data, imbalance_ratio):
     num_samples_0 = train_data.targets.count(0)
     num_classes = float(len(train_data.classes))
     imbalance_factor = math.exp(-math.log(imbalance_ratio)/(num_classes-1))
+    pi_list = []
     all_indexes_seperate = []
     for i in range(10): # 10 
         i_indexes = [k for k, j in enumerate(train_data.targets) if j == i]
         num_samples_i = int(num_samples_0 * (imbalance_factor)**(i))
         sample_indexes = random.sample(i_indexes, num_samples_i)
         all_indexes_seperate.append(sample_indexes)
+        pi_list.append(len(sample_indexes))
+
         
     all_indexes = sum(all_indexes_seperate, []) # collapse list
+    pi_list_tensor = torch.tensor(pi_list) / len(all_indexes)
+
     sampled_images = train_data.data[all_indexes] 
     sampled_targets = np.array(train_data.targets)[all_indexes]
     train_data.data = sampled_images
     train_data.targets = list(sampled_targets)
-    return train_data
+    return train_data, pi_list_tensor
+    
 
 
 def get_loaders(train_data, test_data):
@@ -85,11 +92,23 @@ def test(model, device, testloader):
   preds = np.array(preds).reshape(-1)
   cm = confusion_matrix(list(truths), list(preds))
   wandb.log({'Test Accuracy': 100 * correct / total})
+  print('Accuracy: ',100*correct/total)
   print('Class-wise accuracy:',cm.diagonal()/cm.sum(axis=1))
   print('f1 score: ',f1_score(truths, preds, average=None))
 
 
-# def finetune(model, trainloader, criterion, optimizer, scheduler, device, testloader):
+def get_balanced_loader(train_data, pi_tensor):
+  y_train = [train_data.targets[i] for i in np.arange(len(train_data.targets))]
+  b = []
+  for t in y_train:
+    b.append(pi_tensor[t])
+
+  sample_wts = np.array(b)
+
+  sampler = torch.utils.data.WeightedRandomSampler(weights= 1/sample_wts, num_samples = len(b), replacement = True)
+  balanced_loader = torch.utils.data.DataLoader(train_data, batch_size=config.batch_size, sampler=sampler,  num_workers=8)
+  return balanced_loader
+
 
 
 
@@ -108,7 +127,7 @@ def main():
 
     cifar10_train = datasets.CIFAR10('/workspace/Datasets/CIFAR10/train', download=False, train=True, transform=transform_train)
     cifar10_test = datasets.CIFAR10('/workspace/Datasets/CIFAR10/test',download=False, train=False, transform=transform_test)
-    cifar10_train = make_lt_dataset(cifar10_train, 100)
+    cifar10_train, pi_tensor = make_lt_dataset(cifar10_train, 100)
     trainloader, testloader = get_loaders(cifar10_train, cifar10_test)
     model = resnet32()
     criterion = nn.CrossEntropyLoss()
@@ -116,21 +135,65 @@ def main():
     scheduler = torch.optim.lr_scheduler.MultiStepLR(optimizer, milestones = [150, 180], gamma = 0.1)
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
     wandb.watch(model, log='all')
-    print("Joint Training")
-    print("*************************************************")
-    train(model, trainloader, criterion, optimizer, scheduler, device, testloader, config.num_epochs)
-    
-    
-    torch.save(model, 'checkpoint')
+
+    if(path.exists('checkpoint')):
+      model.load_state_dict(torch.load('checkpoint'))
+      print('model loaded')
+
+    else:
+      print("Joint Training")
+      print("*************************************************")
+      train(model, trainloader, criterion, optimizer, scheduler, device, testloader, config.num_epochs)
+          
+      torch.save(model, 'checkpoint')
+      
     for param in model.parameters():
       param.requires_grad = False
+    
     model.linear = nn.Linear(64, 10)
     print("Finetuning")
     print("*************************************************")
-    train(model, trainloader, criterion, optimizer, scheduler, device, testloader, config.num_epochs_fine)
+    finetune_loader = get_balanced_loader(cifar10_train, pi_tensor)
+    train(model, finetune_loader, criterion, optimizer, scheduler, device, testloader, config.num_epochs_fine)
 
     
 if (__name__=="__main__"):
   # main()
-  model = torch.load('checkpoint')
+  transform_train = transforms.Compose([
+transforms.RandomCrop(32, padding=4),
+transforms.RandomHorizontalFlip(),
+transforms.ToTensor(),
+transforms.Normalize((0.4914, 0.4822, 0.4465), (0.2023, 0.1994, 0.2010)),
+])
+  transform_test = transforms.Compose([
+transforms.ToTensor(),
+transforms.Normalize((0.4914, 0.4822, 0.4465), (0.2023, 0.1994, 0.2010)),
+])
+  model = models.resnet34(pretrained=True)
+
+  for param in model.parameters():
+      param.requires_grad = False
+  
+  n = model.fc.in_features
+
+  class Identity(nn.Module):
+    def __init__(self):
+       super(Identity, self).__init__()
+
+    def forward(self, x):
+      return x
+  model.fc = nn.Linear(n, 10)
+  model.avgpool = Identity()
+
   print(model)
+  cifar10_train = datasets.CIFAR10('/workspace/Datasets/CIFAR10/train', download=False, train=True, transform=transform_train)
+  cifar10_test = datasets.CIFAR10('/workspace/Datasets/CIFAR10/test',download=False, train=False, transform=transform_test)
+  cifar10_train, pi_tensor = make_lt_dataset(cifar10_train, 100)
+  trainloader, testloader = get_loaders(cifar10_train, cifar10_test)
+  finetune_loader = get_balanced_loader(cifar10_train, pi_tensor)
+  criterion = nn.CrossEntropyLoss()
+  optimizer = torch.optim.SGD(model.parameters(), lr=config.lr, momentum = config.momentum, weight_decay=config.weight_decay)
+  scheduler = torch.optim.lr_scheduler.MultiStepLR(optimizer, milestones = [150, 180], gamma = 0.1)
+  device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+  train(model, finetune_loader, criterion, optimizer, scheduler, device, testloader, 100)
+
